refactor: refactoring pull-request-compare-page
fix: minor changes

fix: minor changes

fix: minor changes

fix: minor changes

fix: minor changes

fix: minor changes
