+++ apps/gitness/src/pages-v2/pull-request/pull-request-compare.tsx
@@ -25,7 +25,7 @@ import {
   BranchSelectorTab,
   CommitSelectorListItem,
   CompareFormFields,
-  PullRequestCompare
+  PullRequestComparePage
 } from '@harnessio/ui/views'
 
 import { useAppContext } from '../../framework/context/AppContext'
@@ -365,7 +365,7 @@ export const CreatePullRequest = () => {
     if (isFetchingBranches) return <SkeletonList />
 
     return (
-      <PullRequestCompare
+      <PullRequestComparePage
         currentUser={currentUser?.display_name}
         setSearchCommitQuery={setQuery}
         searchCommitQuery={query}
