+++ packages/ui/src/views/index.ts
@@ -5,7 +5,6 @@ export * from './types/link-types'
 export * as SandboxLayout from './layouts/SandboxLayout'
 export * from './layouts/SandboxRoot'
 export * from './layouts/Floating1ColumnLayout'
-export * from './layouts/PullRequestCompareLayout'
 export * from './layouts/PullRequestLayout'
 
 // Account
