+++ packages/ui/src/views/repo/pull-request/compare/components/pull-request-compare-form.tsx
@@ -2,7 +2,6 @@ import { forwardRef } from 'react'
 import { FieldErrors, SubmitHandler, UseFormHandleSubmit, UseFormRegister } from 'react-hook-form'
 
 import { Fieldset, Input, Textarea } from '@/components'
-import { Text } from '@components/text'
 import { z } from 'zod'
 
 // Define the form schema
@@ -31,28 +30,27 @@ const PullRequestCompareForm = forwardRef<HTMLFormElement, PullRequestFormProps>
     }
     return (
       <form ref={ref} onSubmit={handleSubmit(onSubmit)}>
-        <Fieldset>
+        <Fieldset className="gap-y-3">
+          <span className="text-18 font-medium leading-none text-foreground-1">Add a title</span>
           <Input
             id="title"
             {...register('title')}
             placeholder="Enter pull request title"
-            label="Title"
             error={errors.title?.message?.toString()}
             autoFocus
+            size="md"
           />
+          <span className="mt-4 text-18 font-medium leading-none text-foreground-1">Add a description</span>
           <Textarea
             id="description"
             {...register('description')}
             placeholder="Add Pull Request description here."
-            label="Description"
             error={errors.description?.message?.toString()}
           />
         </Fieldset>
 
         {apiError && apiError !== "head branch doesn't contain any new commits." && (
-          <Text size={1} className="text-destructive">
-            {apiError?.toString()}
-          </Text>
+          <span className="text-destructive text-12">{apiError?.toString()}</span>
         )}
       </form>
     )
