+++ packages/ui/src/views/repo/pull-request/compare/components/pull-request-compare-tab-trigger-item.tsx
@@ -1,3 +1,5 @@
+import { FC } from 'react'
+
 import { Badge } from '@components/badge'
 import { Icon, IconProps } from '@components/icon'
 import { TabsTrigger } from '@components/tabs'
@@ -11,9 +13,9 @@ interface TabTriggerItemProps {
   badgeCount?: number // Optional badge count
 }
 
-const TabTriggerItem: React.FC<TabTriggerItemProps> = ({ value, icon, label, badgeCount }) => {
+const TabTriggerItem: FC<TabTriggerItemProps> = ({ value, icon, label, badgeCount }) => {
   return (
-    <TabsTrigger value={value}>
+    <TabsTrigger value={value} className="data-[state=active]:bg-background-1">
       <Layout.Horizontal className="items-center">
         <div>
           <Icon size={16} name={icon as IconProps['name']} />
