--- packages/ui/src/views/repo/pull-request/compare/components/types.ts
@@ -1,4 +1,0 @@
-export interface CommitSelectorListItem {
-  title: string
-  sha: string
-}
