+++ packages/ui/src/views/repo/pull-request/compare/pull-request-compare.types.ts
@@ -1,0 +1,11 @@
+export interface HeaderProps {
+  text: string
+  data?: string
+  title: string
+  lang: string
+  addedLines?: number
+  removedLines?: number
+  isBinary?: boolean
+  deleted?: boolean
+  unchangedPercentage?: number
+}
