+++ packages/ui/src/views/repo/pull-request/components/pull-request-diff-viewer.tsx
@@ -7,8 +7,8 @@ import { DiffBlock } from 'diff2html/lib/types'
 import { debounce } from 'lodash-es'
 import { OverlayScrollbars } from 'overlayscrollbars'
 
+import constants from '../constants'
 import { useDiffHighlighter } from '../hooks/useDiffHighlighter'
-import constants from './constants'
 
 const TextArea = ({ onChange }: { onChange: (v: string) => void }) => {
   const [val, setVal] = useState('')
