+++ packages/ui/src/views/repo/pull-request/details/components/conversation/pull-request-overview.tsx
@@ -5,7 +5,7 @@ import { DiffModeEnum } from '@git-diff-view/react'
 import { getInitials } from '@utils/stringUtils'
 import { timeAgo } from '@utils/utils'
 import { TranslationStore } from '@views/index'
-import PullRequestDiffViewer from '@views/repo/pull-request/diff-viewer/pull-request-diff-viewer'
+import PullRequestDiffViewer from '@views/repo/pull-request/components/pull-request-diff-viewer'
 import { useDiffConfig } from '@views/repo/pull-request/hooks/useDiffConfig'
 import { TypesPullReq } from '@views/repo/pull-request/pull-request.types'
 import { parseStartingLineIfOne } from '@views/repo/pull-request/utils'
