+++ packages/ui/src/views/repo/pull-request/index.ts
@@ -1,6 +1,8 @@
 //pull-request-list
 export * from './pull-request-list-page'
-export * from './compare/components/types'
+export * from './pull-request.types'
+//pull-request-compare
+export * from './compare/pull-request-compare-page'
 //pull-request-commits
 export * from './commits/pull-request-commits'
 export * from './commits/pull-request-commits.types'
@@ -11,7 +13,7 @@ export * from './details/pull-request-details-types'
 export * from './details/pull-request-changes-page'
 
 //pull-request-conversation
-export * from './details/components/conversation/pull-request-side-bar'
+export * from './components/pull-request-side-bar'
 export * from './details/components/conversation/pull-request-comment-box'
 export * from './details/components/conversation/pull-request-overview'
 export * from './details/components/conversation/pull-request-panel'
