+++ packages/ui/src/views/repo/repo-commits/index.tsx
@@ -1,4 +1,4 @@
-export * from './components/commits-list'
+export * from '../components/commits-list'
 export * from './components/filter'
 export * from './types'
 export * from './repo-commits-view'
