+++ packages/ui/src/views/repo/repo-commits/types.ts
@@ -1,42 +1,4 @@
-export interface TypesCommit {
-  author?: TypesSignature
-  committer?: TypesSignature
-  message?: string
-  parent_shas?: string[]
-  sha?: string
-  stats?: TypesCommitStats
-  title?: string
-}
-
-export interface TypesSignature {
-  identity?: TypesIdentity
-  when?: string
-}
-
-export interface TypesCommitStats {
-  files?: TypesCommitFileStats[]
-  total?: TypesChangeStats
-}
-
-export interface TypesIdentity {
-  email?: string
-  name?: string
-}
-
-export interface TypesCommitFileStats {
-  changes?: number
-  deletions?: number
-  insertions?: number
-  old_path?: string
-  path?: string
-  status?: string
-}
-
-export interface TypesChangeStats {
-  changes?: number
-  deletions?: number
-  insertions?: number
-}
+import { TypesCommit } from '@/views'
 
 export interface TypesListCommitResponse {
   commits?: TypesCommit[] | null
