+++ packages/ui/src/views/repo/repo.types.ts
@@ -114,3 +114,42 @@ export enum CodeModes {
   NEW = 'new',
   VIEW = 'view'
 }
+
+export interface TypesChangeStats {
+  changes?: number
+  deletions?: number
+  insertions?: number
+}
+
+export interface TypesSignature {
+  identity?: TypesIdentity
+  when?: string
+}
+
+export interface TypesIdentity {
+  email?: string
+  name?: string
+}
+
+export interface TypesCommit {
+  author?: TypesSignature
+  committer?: TypesSignature
+  message?: string
+  parent_shas?: string[]
+  sha?: string
+  stats?: TypesCommitStats
+  title?: string
+}
+export interface TypesCommitStats {
+  files?: TypesCommitFileStats[]
+  total?: TypesChangeStats
+}
+
+export interface TypesCommitFileStats {
+  changes?: number
+  deletions?: number
+  insertions?: number
+  old_path?: string
+  path?: string
+  status?: string
+}
